refactor: breaking change `decode()` now return `Lnurl` (#49)
returning an Union type here is a big mistake, because its tricky to
access the actual useful Lnurl class.

there is a chance it does not break anthing because usually it was used
like: `str(url)` because it was of `Url` type anyway, and __str__ of
Lnurl should give the same results.